requirement.txt
asgiref==3.3.4
boto3==1.17.74
botocore==1.20.74
Django==3.2.3
django-storages==1.11.1
djangorestframework==3.12.4
jmespath==0.10.0
Pillow==8.2.0
pymemcache==3.4.3
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
